requirements.txt
# Core
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.5.0,<3.0.0

# Database
sqlalchemy[asyncio]>=2.0.35,<3.0.0
asyncpg>=0.30.0,<1.0.0
alembic>=1.14.0,<2.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=3.2.0,<4.0.0
python-multipart>=0.0.12

# HTTP client (Connector Gateway external calls)
httpx>=0.27.0,<1.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
